[Refactor]: Extract hub profile enpoints to own file
