+++ cardstack/src/services/hub/hub-api.ts
@@ -1,19 +1,12 @@
 import { NativeCurrency } from '@cardstack/cardpay-sdk';
 import { createApi } from '@reduxjs/toolkit/query/react';
 
-import { CustodialWallet, ProfileIDUniquenessResponse } from '@cardstack/types';
+import { CustodialWallet } from '@cardstack/types';
 import { transformObjKeysToCamelCase } from '@cardstack/utils';
 
-import { CacheTags as SafesCacheTags, safesApi } from '../safes-api';
 import { queryPromiseWrapper } from '../utils';
 
-import {
-  checkHubAuth,
-  fetchHubBaseQuery,
-  hubBodyBuilder,
-  hubProfilePurchaseBody,
-  patchProfileInfo,
-} from './hub-service';
+import { checkHubAuth, fetchHubBaseQuery, hubBodyBuilder } from './hub-service';
 import {
   GetCustodialWalletQueryResult,
   RequestCardDropQueryParams,
@@ -23,13 +16,6 @@ import {
   CheckHubAuthQueryParams,
   RegisterFCMTokenQueryParams,
   GetExchangeRatesQueryParams,
-  PostProfilePurchaseQueryParams,
-  GetValidateProfileSlugParams,
-  CreateProfileInfoParams,
-  PostProfilePurchaseQueryResult,
-  UpdateProfileInfoParams,
-  JobTicketTypeResult,
-  JobStateType,
 } from './hub-types';
 
 const routes = {
@@ -37,12 +23,6 @@ const routes = {
   emailDrop: '/email-card-drop-requests',
   exchangeRates: '/exchange-rates',
   registerFCMToken: '/push-notification-registrations',
-  profilePurchases: '/profile-purchases',
-  profileInfo: {
-    root: '/profiles',
-    validateSlug: '/validate-slug',
-    jobTicket: '/job-tickets',
-  },
 };
 
 enum CacheTag {
@@ -117,75 +97,6 @@ export const hubApi = createApi({
         responseHandler: response => response.text(),
       }),
     }),
-    profilePurchases: builder.mutation<string, PostProfilePurchaseQueryParams>({
-      query: params => ({
-        url: routes.profilePurchases,
-        method: 'POST',
-        body: hubProfilePurchaseBody(routes.profilePurchases, params),
-      }),
-      transformResponse: (result: PostProfilePurchaseQueryResult) =>
-        result?.included?.[0].id,
-    }),
-    validateProfileSlug: builder.query<
-      ProfileIDUniquenessResponse,
-      GetValidateProfileSlugParams
-    >({
-      query: ({ slug }) =>
-        `${routes.profileInfo.root}${routes.profileInfo.validateSlug}/${slug}`,
-    }),
-    createProfileInfo: builder.mutation<string, CreateProfileInfoParams>({
-      query: params => ({
-        url: routes.profileInfo.root,
-        method: 'POST',
-        body: hubBodyBuilder(routes.profileInfo.root, params),
-      }),
-      transformResponse: ({ data }) => data?.attributes?.did,
-    }),
-    updateProfileInfo: builder.mutation<string, UpdateProfileInfoParams>({
-      query: params => ({
-        url: `${routes.profileInfo.root}/${params.id}`,
-        method: 'PATCH',
-        body: hubBodyBuilder(routes.profileInfo.root, params),
-      }),
-      async onQueryStarted(params, { queryFulfilled }) {
-        const updatedResult = patchProfileInfo(params);
-
-        queryFulfilled.catch(updatedResult.undo);
-      },
-    }),
-    getProfileJobStatus: builder.query<JobStateType, { jobTicketID: string }>({
-      query: ({ jobTicketID }) =>
-        `${routes.profileInfo.jobTicket}/${jobTicketID}`,
-      transformResponse: ({ data }: { data: JobTicketTypeResult }) =>
-        data.attributes?.state,
-      async onQueryStarted(_, { dispatch, queryFulfilled }) {
-        try {
-          const { data: status } = await queryFulfilled;
-
-          if (status === 'success') {
-            dispatch(safesApi.util.invalidateTags([SafesCacheTags.SAFES]));
-          }
-        } catch {}
-      },
-    }),
-    postProfileJobRetry: builder.mutation<unknown, { jobTicketID: string }>({
-      query: ({ jobTicketID }) => ({
-        url: `${routes.profileInfo.jobTicket}/${jobTicketID}/retry`,
-        method: 'POST',
-      }),
-    }),
-    getProfileUnfulfilledJob: builder.query<string | undefined, void>({
-      query: () => `${routes.profileInfo.jobTicket}`,
-      transformResponse: ({ data }: { data: JobTicketTypeResult[] }) => {
-        const unfulfilledJob = data.find(
-          ({ attributes }) =>
-            attributes['job-type'] === 'create-profile' &&
-            attributes.state !== 'success'
-        );
-
-        return unfulfilledJob?.id;
-      },
-    }),
   }),
 });
 
@@ -197,11 +108,4 @@ export const {
   useGetExchangeRatesQuery,
   useRegisterFcmTokenQuery,
   useUnregisterFcmTokenMutation,
-  useProfilePurchasesMutation,
-  useLazyValidateProfileSlugQuery,
-  useCreateProfileInfoMutation,
-  useGetProfileJobStatusQuery,
-  usePostProfileJobRetryMutation,
-  useUpdateProfileInfoMutation,
-  useGetProfileUnfulfilledJobQuery,
 } = hubApi;
