+++ cardstack/src/services/index.ts
@@ -8,3 +8,4 @@ export * from './service-status-api';
 export * from './merchant/merchant-api';
 export * from './prepaid-cards/prepaid-card-api';
 export * from './opensea-api';
+export * from './hub/endpoints/hub-profile-api';
